Full refactoring models path on all project
