+++ app/Http/Controllers/HomeController.php
@@ -1,34 +1,20 @@
 <?php
 
 namespace App\Http\Controllers;
-//namespace App\Http\Model;
-
-use Illuminate\Http\Request;
-use App\Http\Model\HomeModel;
-
-//use App\HomeModel;
+use App\Models\HomeModel;
 
 class HomeController extends Controller
 {
-    /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-       // $this->middleware('auth');
-    }
 
     /**
-     * Show the application dashboard.
+     * Show the home page
      *
-     * @return \Illuminate\Contracts\Support\Renderable
+     * @return
      */
     public function index()
     {
        $notations = HomeModel::take_notations();
-      
+
         return view('home', compact('notations'));
     }
 }
