+++ app/Http/Controllers/MapController.php
@@ -2,10 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-use App\Http\Model\NotationModel;
-use App\Http\Requests\NotationRequest;
-use Illuminate\Support\Facades\Auth;
+use App\Models\NotationModel;
 
 class MapController extends Controller
 {
