+++ app/Http/Controllers/NotationController.php
@@ -3,8 +3,8 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Http\Model\NotationModel;
-use App\Http\Model\Notation\NotationViewModel;
+use App\Models\NotationModel;
+use App\Models\Notation\NotationViewModel;
 use App\Http\Requests\NotationRequest;
 use App\Http\Requests\NotationPhotoRequest;
 use Illuminate\Support\Facades\Auth;
@@ -16,7 +16,7 @@ class NotationController extends Controller
 
         $input = $request->only(['name_tema','text_notation', 'method']);
         $data = NotationModel::createNotation($input);
-    
+
         return response()->json(['notationData' => $data]);
     }
 
@@ -26,20 +26,20 @@ class NotationController extends Controller
         {
             NotationViewModel::addViewNotation($notationId);
             $view = NotationModel::viewNotation($notationId);
-                   
+
         }
-        catch (\Exception $exception) 
+        catch (\Exception $exception)
         {
             return view('error_404', ['error' => ['Данной статьи не существует']]);
         }
 
         return view('menu.notation_view', ['view' => $view]);
-      
+
     }
 
     protected function NotationEditAccess(int $notationId)
     {
-        try 
+        try
         {
             $data_edit = NotationModel::dataEditNotation($notationId);
 
@@ -47,8 +47,8 @@ class NotationController extends Controller
                 return view('menu.Notation.notation_edit', ['data_notation' => $data_edit['notation'],
                         'photo_notation' => $data_edit['notation_photos']]);
            else  return view('error_404', ['error' => ['Доступ на редактирование запрещён']]);
-        } 
-        catch (\Exception $exception) 
+        }
+        catch (\Exception $exception)
         {
             return view('error_404', ['error' => ['Данной статьи не существует']]);
         }
@@ -69,7 +69,7 @@ class NotationController extends Controller
     {
         if($request->ajax())
         {
-            $input = $request->all(); 
+            $input = $request->all();
             $back = NotationModel::notationRating($input['notation_id'], $input['action']);
 
             return response()->json(['success'=> $back]);
@@ -82,14 +82,14 @@ class NotationController extends Controller
         {
             $input = $request->only(['notation_id']); //получение всех входных данных
             $back = NotationModel::notationDelete($input['notation_id']);
-           
+
             return response()->json(['success'=> $back]);
         }
     }
 
     //NotationPhotoRequest Request
     protected function NotationAddPhotos(NotationPhotoRequest $request)
-    { 
+    {
         $paths = NotationModel::notationAddPhotos($request);
 
         if(!empty($paths))
@@ -98,18 +98,18 @@ class NotationController extends Controller
             ->with('success', "Изображения загружены успешно.")
             ->with('paths', $paths);
         }
-        else 
+        else
             return back()->with('error', "Изображения не загружены!");
     }
 
     protected function NotationPhotoDelete(Request $request)
     {
-        $photo_data = $request->only(['photo_id', 'notation_id']); 
+        $photo_data = $request->only(['photo_id', 'notation_id']);
         $del = NotationModel::notationPhotoDelete($photo_data);
 
         return response()->json(['answer'=> $del]);
     }
-    
+
     public function Notation(Request $request)
     {
         return view('menu.notation');
