+++ app/Http/Controllers/ProfileController.php
@@ -3,21 +3,21 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Http\Model\ProfileModel;
+use App\Models\ProfileModel;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\ProfileAvatarRequest;
 use Validator;
 
 class ProfileController extends Controller
 {
-    
+
     public function viewAnotherProfile(int $id)
     {
         $userData = ProfileModel::getAnotherUser($id);
-            
+
         if(!empty($userData->name))
             return view('menu.profile', ['data_user' => $userData]);
-        else  
+        else
             return view('error_404', ['error' => ['Данного пользователя не существует']]);
     }
 
@@ -39,9 +39,9 @@ class ProfileController extends Controller
     {
         if($request->ajax())
         {
-            $input = $request->only(['data_send']); 
+            $input = $request->only(['data_send']);
             $back = ProfileModel::changeProfile($input);
-                
+
             try
             {
                 if(!$back->original['status'])
@@ -61,10 +61,10 @@ class ProfileController extends Controller
     public function changeAvatar(ProfileAvatarRequest $request)
     {
         $isChanged = ProfileModel::changeAvatar($request);
-   
+
         return redirect()->route('profile_id', Auth::user()->id)
             ->with('success','Вы успешно изменили аватар')
             ->with('isChanged', $isChanged);
-           
+
     }
 }
