--- app/Http/Model/Api/v1/ApiNotationModel.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Http\Model\Api\v1;
-
-use Illuminate\Database\Eloquent\Model;
-
-class ApiNotationModel extends Model
-{
-    //
-}
